Convert some endpoints to async/await

+++ server/server.js
@@ -29,7 +29,7 @@ app.post('/todos', authenticate, (req, res) => {
         .catch((error) => res.status(400).send(error));
 });
 
-app.get('/todos', authenticate , (req, res) => {
+app.get('/todos', authenticate, (req, res) => {
     Todo.find({
         _creator: req.user._id
     })
@@ -56,23 +56,20 @@ app.get('/todos/:id', authenticate, (req, res) => {
         .catch((error) => res.status(400).send());
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+app.delete('/todos/:id', authenticate, async (req, res) => {
     const { id } = req.params;
 
     if (!ObjectID.isValid(id))
         return res.status(404).send();
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    })
-        .then((todo) => {
-            if (!todo)
-                return res.status(404).send();
-
-            res.send({ todo });
-        })
-        .catch((error) => res.status(400).send());
+    try {
+        const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
+        if (!todo)
+            return res.status(404).send();
+        res.send({ todo });
+    } catch (error) {
+        res.status(400).send()
+    }
 });
 
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -84,7 +81,7 @@ app.patch('/todos/:id', authenticate, (req, res) => {
 
     body.completedAt = (_.isBoolean(body.completed) && body.completed) ? new Date().getTime() : null;
 
-    Todo.findOneAndUpdate({ _id: id, _creator: req.user._id } , { $set: body }, { new: true })
+    Todo.findOneAndUpdate({ _id: id, _creator: req.user._id }, { $set: body }, { new: true })
         .then((todo) => {
             if (!todo)
                 return res.status(404).send();
@@ -94,37 +91,43 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         .catch((error) => res.status(400).send())
 });
 
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
     const body = _.pick(req.body, ['email', 'password']);
     const user = new User(body);
-    user.save()
-        .then(() => user.generateAuthToken())
-        .then((token) => res.header('x-auth', token).send(user))
-        .catch((error) => res.status(400).send({ error }));
+
+    try {
+        await user.save();
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (error) {
+        res.status(400).send({ error });
+    }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
+app.post('/users/login', async (req, res) => {
     const { email, password } = req.body;
 
-    User.findByCredentials(email, password)
-        .then((user) => {
-            return user.generateAuthToken()
-                .then((token) => {
-                    res.header('x-auth', token).send({ user });
-                });
-        })
-        .catch((error) => res.status(400).send());
-
+    try {
+        const user = await User.findByCredentials(email, password);
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send({ user });
+    } catch (error) {
+        res.status(400).send()
+    }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token)
-        .then(() => res.send())
-        .catch(() => res.status(400).send());
+app.delete('/users/me/token', authenticate, async (req, res) => {
+
+    try {
+        await req.user.removeToken(req.token);
+        res.send();
+    } catch (error) {
+        res.status(400).send()
+    }
 });
 
 app.listen(port, () => console.log(`Listening on port ${port}...`));
